// webpack---sb/playerui/src/platformui/Store/Bonuses/Model/Enums/EBonusHelperButton.ts
enum EBonusHelperButton {
  deposit = "DEPOSIT",
  betting = "BETTING",
  liveBetting = "LIVE_BETTING",
  casino = "CASINO",
  casinoProvider = "CASINO_PROVIDER",
  liveCasino = "LIVE_CASINO",
  liveCasinoProvider = "LIVE_CASINO_PROVIDER",
  games = "GAMES",
  gamesProvider = "GAMES_PROVIDER",
  casinoFreeSpins = "CASINO_FREE_SPINS",
}

export { EBonusHelperButton };